fix 'grasshopper grade book' in C#

+++ codewars/csharp/kyu8/grasshopper_grade_book/grasshopper_grade_book.cs
@@ -4,7 +4,7 @@ public class GrasshopperGradeBook
 {
     public static char Solution(int s1, int s2, int s3)
     {
-        return ((double)(s1 + s2 + s3) / 30) switch
+        return ((s1 + s2 + s3) / 30) switch
         {
             10 => 'A',
             9 => 'A',
